+++ modules/tinymce/src/core/main/ts/api/EventTypes.ts
@@ -58,6 +58,17 @@ export interface LoadErrorEvent { message: string }
 export interface PreProcessEvent extends ParserArgs { node: Element }
 export interface PostProcessEvent extends ParserArgs { content: string }
 
+export interface NewTableRowEvent { node: HTMLTableRowElement }
+export interface NewTableCellEvent { node: HTMLTableCellElement }
+
+export interface TableEventData {
+  readonly structure: boolean;
+  readonly style: boolean;
+}
+export interface TableModifiedEvent extends TableEventData {
+  readonly table: HTMLTableElement;
+}
+
 export interface EditorEventMap extends Omit<NativeEventMap, 'blur' | 'focus'> {
   'activate': { relatedTarget: Editor };
   'deactivate': { relatedTarget: Editor };
@@ -122,6 +133,9 @@ export interface EditorEventMap extends Omit<NativeEventMap, 'blur' | 'focus'> {
   'AutocompleterStart': AutocompleterEventArgs;
   'AutocompleterUpdate': AutocompleterEventArgs;
   'AutocompleterEnd': { };
+  'TableModified': TableModifiedEvent;
+  'NewRow': NewTableRowEvent;
+  'NewCell': NewTableCellEvent;
 }
 
 export interface EditorManagerEventMap {