+++ modules/tinymce/src/core/main/ts/api/dom/Selection.ts
@@ -5,7 +5,8 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Type } from '@ephox/katamari';
+import { Selections, SelectionTypes } from '@ephox/darwin';
+import { Arr, Fun, Type } from '@ephox/katamari';
 import { Compare, SugarElement } from '@ephox/sugar';
 
 import { Bookmark } from '../../bookmark/BookmarkTypes';
@@ -24,6 +25,8 @@ import * as NormalizeRange from '../../selection/NormalizeRange';
 import * as SelectionBookmark from '../../selection/SelectionBookmark';
 import { hasAnyRanges, moveEndPoint } from '../../selection/SelectionUtils';
 import * as SetSelectionContent from '../../selection/SetSelectionContent';
+import { ephemera as cellEphemera } from '../../table/TableEphemera';
+import * as TableSelection from '../../table/TableSelection';
 import Editor from '../Editor';
 import AstNode from '../html/Node';
 import BookmarkManager from './BookmarkManager';
@@ -31,6 +34,8 @@ import ControlSelection from './ControlSelection';
 import DOMUtils from './DOMUtils';
 import SelectorChanged from './SelectorChanged';
 import DomSerializer from './Serializer';
+import { TableCellSelection } from './TableCellSelection';
+import { TableResizeHandler } from './TableResizeHandler';
 
 /**
  * This class handles text and control selection it's an crossbrowser utility class.
@@ -57,6 +62,8 @@ const isValidRange = (rng: Range) => {
 interface EditorSelection {
   bookmarkManager: BookmarkManager;
   controlSelection: ControlSelection;
+  _tableResizeHandler: TableResizeHandler;
+  _tableCellSelection: TableCellSelection;
   dom: DOMUtils;
   win: Window;
   serializer: DomSerializer;
@@ -84,6 +91,7 @@ interface EditorSelection {
   getStart: (real?: boolean) => Element;
   getEnd: (real?: boolean) => Element;
   getSelectedBlocks: (startElm?: Element, endElm?: Element) => Element[];
+  getSelectedCells: () => HTMLTableCellElement[];
   normalize: () => Range;
   selectorChanged: (selector: string, callback: (active: boolean, args: {
     node: Node;
@@ -102,6 +110,9 @@ interface EditorSelection {
   destroy: () => void;
 }
 
+const isRoot = (editor: Editor) => (element: SugarElement<Node>): boolean =>
+  Compare.eq(element, SugarElement.fromDom(editor.getBody()));
+
 /**
  * Constructs a new selection instance.
  *
@@ -115,6 +126,15 @@ interface EditorSelection {
 const EditorSelection = (dom: DOMUtils, win: Window, serializer: DomSerializer, editor: Editor): EditorSelection => {
   let selectedRange: Range | null;
   let explicitRange: Range | null;
+
+  const cellSelection = Selections(
+    () => SugarElement.fromDom(editor.getBody()),
+    () => TableSelection.getSelectionCell(SugarElement.fromDom(getStart()), isRoot(editor)),
+    cellEphemera.selectedSelector
+  );
+
+  const _tableResizeHandler = TableResizeHandler(editor);
+  const _tableCellSelection = TableCellSelection(editor);
 
   const { selectorChangedWithUnbind } = SelectorChanged(dom, editor);
 
@@ -451,7 +471,21 @@ const EditorSelection = (dom: DOMUtils, win: Window, serializer: DomSerializer, 
    */
   const getNode = (): Element => ElementSelection.getNode(editor.getBody(), getRng());
 
-  const getSelectedBlocks = (startElm: Element, endElm: Element) => ElementSelection.getSelectedBlocks(dom, getRng(), startElm, endElm);
+  const getSelectedBlocks = (startElm: Element, endElm: Element) =>
+    ElementSelection.getSelectedBlocks(dom, getRng(), startElm, endElm);
+
+  // TODO: TINY-8386 Try and come up with a better and more integrated solution than this
+  const getSelectedCells = (): HTMLTableCellElement[] =>
+    SelectionTypes.fold<HTMLTableCellElement[]>(cellSelection.get(),
+      // No fake selected cells
+      Fun.constant([]),
+      // This path is taken whenever there is fake cell selection even for just a single selected cell
+      (cells) => {
+        return Arr.map(cells, (cell) => cell.dom);
+      },
+      // For this path, the start of the selection whether collapsed or ranged is within a table cell
+      (cell) => [ cell.dom ]
+    );
 
   const isForward = (): boolean => {
     const sel = getSel();
@@ -540,11 +574,14 @@ const EditorSelection = (dom: DOMUtils, win: Window, serializer: DomSerializer, 
   const destroy = () => {
     win = selectedRange = explicitRange = null;
     controlSelection.destroy();
+    _tableResizeHandler.destroy();
   };
 
   const exports: EditorSelection = {
     bookmarkManager: null,
     controlSelection: null,
+    _tableResizeHandler,
+    _tableCellSelection,
     dom,
     win,
     serializer,
@@ -566,6 +603,7 @@ const EditorSelection = (dom: DOMUtils, win: Window, serializer: DomSerializer, 
     getStart,
     getEnd,
     getSelectedBlocks,
+    getSelectedCells,
     normalize,
     selectorChanged,
     selectorChangedWithUnbind,