+++ modules/tinymce/src/core/main/ts/table/InsertTable.ts
@@ -6,15 +6,13 @@
  */
 
 import { Arr, Fun, Type } from '@ephox/katamari';
-import { TableRender } from '@ephox/snooker';
+import { TableRender, TableConversions } from '@ephox/snooker';
 import { Attribute, Html, SelectorFilter, SelectorFind, SugarElement } from '@ephox/sugar';
 
-import Editor from 'tinymce/core/api/Editor';
-
-import { fireNewCell, fireNewRow } from '../api/Events';
+import Editor from '../api/Editor';
 import * as Options from '../api/Options';
-import * as Util from '../core/Util';
-import { enforceNone, enforcePercentage, enforcePixels } from './EnforceUnit';
+import * as Events from '../api/TableEvents';
+import * as Utils from './TableUtils';
 
 const placeCaretInCell = (editor: Editor, cell: SugarElement<HTMLTableCellElement>): void => {
   editor.selection.select(cell.dom, true);
@@ -27,10 +25,10 @@ const selectFirstCellInTable = (editor: Editor, tableElm: SugarElement<HTMLTable
 
 const fireEvents = (editor: Editor, table: SugarElement<HTMLTableElement>): void => {
   Arr.each(SelectorFilter.descendants<HTMLTableRowElement>(table, 'tr'), (row) => {
-    fireNewRow(editor, row.dom);
+    Events.fireNewRow(editor, row.dom);
 
     Arr.each(SelectorFilter.descendants<HTMLTableCellElement>(row, 'th,td'), (cell) => {
-      fireNewCell(editor, cell.dom);
+      Events.fireNewCell(editor, cell.dom);
     });
   });
 };
@@ -39,11 +37,11 @@ const isPercentage = (width: string): boolean =>
   Type.isString(width) && width.indexOf('%') !== -1;
 
 const insert = (editor: Editor, columns: number, rows: number, colHeaders: number, rowHeaders: number): HTMLTableElement => {
-  const defaultStyles = Options.getDefaultStyles(editor);
+  const defaultStyles = Options.getTableDefaultStyles(editor);
   const options: TableRender.RenderOptions = {
     styles: defaultStyles,
-    attributes: Options.getDefaultAttributes(editor),
-    colGroups: Options.useColumnGroup(editor)
+    attributes: Options.getTableDefaultAttributes(editor),
+    colGroups: Options.tableUseColumnGroup(editor)
   };
 
   // Don't create an undo level when inserting the base table HTML otherwise we can end up with 2 undo levels
@@ -57,15 +55,15 @@ const insert = (editor: Editor, columns: number, rows: number, colHeaders: numbe
   });
 
   // Enforce the sizing mode of the table
-  return SelectorFind.descendant<HTMLTableElement>(Util.getBody(editor), 'table[data-mce-id="__mce"]').map((table) => {
-    if (Options.isPixelsForced(editor)) {
-      enforcePixels(table);
-    } else if (Options.isResponsiveForced(editor)) {
-      enforceNone(table);
-    } else if (Options.isPercentagesForced(editor) || isPercentage(defaultStyles.width)) {
-      enforcePercentage(table);
+  return SelectorFind.descendant<HTMLTableElement>(Utils.getBody(editor), 'table[data-mce-id="__mce"]').map((table) => {
+    if (Options.isTablePixelsForced(editor)) {
+      TableConversions.convertToPixelSize(table);
+    } else if (Options.isTableResponsiveForced(editor)) {
+      TableConversions.convertToNoneSize(table);
+    } else if (Options.isTablePercentagesForced(editor) || isPercentage(defaultStyles.width)) {
+      TableConversions.convertToPercentSize(table);
     }
-    Util.removeDataStyle(table);
+    Utils.removeDataStyle(table);
     Attribute.remove(table, 'data-mce-id');
     fireEvents(editor, table);
     selectFirstCellInTable(editor, table);
@@ -73,7 +71,7 @@ const insert = (editor: Editor, columns: number, rows: number, colHeaders: numbe
   }).getOr(null);
 };
 
-const insertTableWithDataValidation = (editor: Editor, rows: number, columns: number, options: Record<string, number> = {}, errorMsg: string): HTMLTableElement | null => {
+const insertTable = (editor: Editor, rows: number, columns: number, options: Record<string, number> = {}): HTMLTableElement | null => {
   const checkInput = (val: any) => Type.isNumber(val) && val > 0;
 
   if (checkInput(rows) && checkInput(columns)) {
@@ -82,12 +80,11 @@ const insertTableWithDataValidation = (editor: Editor, rows: number, columns: nu
     return insert(editor, columns, rows, headerColumns, headerRows);
   } else {
     // eslint-disable-next-line no-console
-    console.error(errorMsg);
+    console.error('Invalid values for mceInsertTable - rows and columns values are required to insert a table.');
     return null;
   }
 };
 
 export {
-  insert,
-  insertTableWithDataValidation
+  insertTable
 };