+++ modules/tinymce/src/core/main/ts/table/TableSize.ts
@@ -5,19 +5,24 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
+/*
+ NOTE: This file is duplicated in the following locations:
+  - models/dom/table/queries/TableSize.ts
+ Make sure that if making changes to this file, the other files are updated as well
+ */
+
 import { TableSize } from '@ephox/snooker';
 import { SugarElement } from '@ephox/sugar';
 
-import Editor from 'tinymce/core/api/Editor';
-
+import Editor from '../api/Editor';
 import * as Options from '../api/Options';
 
 export const get = (editor: Editor, table: SugarElement<HTMLTableElement>): TableSize => {
   // Note: We can't enforce none (responsive), as if someone manually resizes a table
   // then it must switch to either pixel (fixed) or percentage (relative) sizing
-  if (Options.isPercentagesForced(editor)) {
+  if (Options.isTablePercentagesForced(editor)) {
     return TableSize.percentageSize(table);
-  } else if (Options.isPixelsForced(editor)) {
+  } else if (Options.isTablePixelsForced(editor)) {
     return TableSize.pixelSize(table);
   } else {
     // Detect based on the table width