+++ modules/tinymce/src/plugins/table/main/ts/Plugin.ts
@@ -5,64 +5,24 @@
  * For commercial licenses see https://www.tiny.cloud/
  */
 
-import { Selections } from '@ephox/darwin';
-
 import Editor from 'tinymce/core/api/Editor';
 import PluginManager from 'tinymce/core/api/PluginManager';
 
-import * as Clipboard from './actions/Clipboard';
-import { getResizeHandler } from './actions/ResizeHandler';
-import { TableActions } from './actions/TableActions';
-import { Api, getApi } from './api/Api';
 import * as Commands from './api/Commands';
 import * as Options from './api/Options';
-import * as QueryCommands from './api/QueryCommands';
-import { Clipboard as FakeClipboard } from './core/Clipboard';
-import * as TableFormats from './core/TableFormats';
-import * as Util from './core/Util';
-import * as TabContext from './queries/TabContext';
-import CellSelection from './selection/CellSelection';
-import { ephemera } from './selection/Ephemera';
 import { getSelectionTargets } from './selection/SelectionTargets';
-import { getSelectionCell } from './selection/TableSelection';
 import * as Buttons from './ui/Buttons';
 import * as MenuItems from './ui/MenuItems';
 
-const Plugin = (editor: Editor): Api => {
+const Plugin = (editor: Editor): void => {
+  const selectionTargets = getSelectionTargets(editor);
+
   Options.register(editor);
+  Commands.registerCommands(editor);
 
-  const selections = Selections(() => Util.getBody(editor), () => getSelectionCell(Util.getSelectionStart(editor), Util.getIsRoot(editor)), ephemera.selectedSelector);
-  const selectionTargets = getSelectionTargets(editor, selections);
-  const resizeHandler = getResizeHandler(editor);
-  const cellSelection = CellSelection(editor, resizeHandler.lazyResize, selectionTargets);
-  const actions = TableActions(editor, cellSelection, resizeHandler.lazyWire);
-  const clipboard = FakeClipboard();
-
-  Commands.registerCommands(editor, actions, cellSelection, selections, clipboard);
-  QueryCommands.registerQueryCommands(editor, actions, selections);
-  Clipboard.registerEvents(editor, selections, actions);
-
-  MenuItems.addMenuItems(editor, selections, selectionTargets, clipboard);
-  Buttons.addButtons(editor, selections, selectionTargets, clipboard);
+  MenuItems.addMenuItems(editor, selectionTargets);
+  Buttons.addButtons(editor, selectionTargets);
   Buttons.addToolbars(editor);
-
-  editor.on('PreInit', () => {
-    editor.serializer.addTempAttr(ephemera.firstSelected);
-    editor.serializer.addTempAttr(ephemera.lastSelected);
-    TableFormats.registerFormats(editor);
-  });
-
-  if (Options.hasTabNavigation(editor)) {
-    editor.on('keydown', (e: KeyboardEvent) => {
-      TabContext.handle(e, editor, cellSelection);
-    });
-  }
-
-  editor.on('remove', () => {
-    resizeHandler.destroy();
-  });
-
-  return getApi(editor, clipboard, resizeHandler, selectionTargets);
 };
 
 export default (): void => {