+++ modules/tinymce/src/plugins/table/test/ts/browser/TabKeyNavigationTest.ts
@@ -4,8 +4,8 @@ import { LegacyUnit, TinyAssertions, TinyContentActions, TinyHooks, TinySelectio
 import { assert } from 'chai';
 
 import Editor from 'tinymce/core/api/Editor';
+import { TableModifiedEvent } from 'tinymce/core/api/EventTypes';
 import { EditorEvent } from 'tinymce/core/api/util/EventDispatcher';
-import { TableModifiedEvent } from 'tinymce/plugins/table/api/Events';
 import Plugin from 'tinymce/plugins/table/Plugin';
 
 describe('browser.tinymce.plugins.table.TabKeyNavigationTest', () => {